appserver/requirements.txt
django == 4.2.2
pillow
django-admin-interface
django-crispy-forms
django-rest-framework